user comments on the timeline
Users can now comment on other users posts using the timeline, will add this to the user profile next

+++ index.php
@@ -12,24 +12,35 @@ $user_id = $row[0];
 $forename = $row[2];
 
 if (isset($_GET['postid'])) {
+
   $postid = $_GET['postid'];
 
-  $sql = "SELECT user_id FROM post_likes WHERE post_id= $postid AND user_id= $user_id";
-  $result = mysqli_query($db, $sql) or die(mysqli_error($db));
-  if (mysqli_num_rows($result) < 1) {
-    $sql = "UPDATE post SET likes = likes + 1 WHERE post = $postid";
+  if (isset($_POST['comment'])) {
+    $commentbody = $_POST['commentbody'];
+    $time = date("Y-m-d H:i:s");
+    $sql = "INSERT INTO users_comments (body, user_id, posted_at, post_id) VALUES ('$commentbody', '$user_id', '$time', '$postid' )";
     $result = mysqli_query($db, $sql) or die(mysqli_error($db));
-    $sql = "INSERT INTO post_likes (post_id, user_id) VALUES ($postid, $user_id)";
-    $result = mysqli_query($db, $sql) or die(mysqli_error($db));
-  } else {
-    $sql = "UPDATE post SET likes = likes - 1 WHERE post = $postid";
-    $result = mysqli_query($db, $sql) or die(mysqli_error($db));
-    $sql = "DELETE FROM post_likes WHERE post_id = $postid AND user_id = $user_id";
-    $result = mysqli_query($db, $sql) or die(mysqli_error($db));
+  }
+  else {
 
+    $sql = "SELECT user_id FROM post_likes WHERE post_id= $postid AND user_id= $user_id";
+    $result = mysqli_query($db, $sql) or die(mysqli_error($db));
+    if (mysqli_num_rows($result) < 1) {
+      $sql = "UPDATE post SET likes = likes + 1 WHERE post = $postid";
+      $result = mysqli_query($db, $sql) or die(mysqli_error($db));
+      $sql = "INSERT INTO post_likes (post_id, user_id) VALUES ($postid, $user_id)";
+      $result = mysqli_query($db, $sql) or die(mysqli_error($db));
+    } else {
+      $sql = "UPDATE post SET likes = likes - 1 WHERE post = $postid";
+      $result = mysqli_query($db, $sql) or die(mysqli_error($db));
+      $sql = "DELETE FROM post_likes WHERE post_id = $postid AND user_id = $user_id";
+      $result = mysqli_query($db, $sql) or die(mysqli_error($db));
+
+    }
   }
 
 }
+
 
 $sql = "SELECT post.post, post.posted_at, post.body, users.username, users.image, post.likes FROM users, post, followers WHERE post.user_id = followers.user_id AND users.user_id = post .user_id AND follower_id = '$user_id' ORDER BY `post`.`posted_at` DESC";
 $result = mysqli_query($db, $sql) or die(mysqli_error($db));
@@ -41,14 +52,30 @@ $posts = "";
 
 while ($row = mysqli_fetch_array($result)) {
   $postid = $row[0];
+  $comments = "";
   $sql = "SELECT post_id FROM post_likes WHERE post_id=$postid and user_id=$user_id";
   $result2 = mysqli_query($db, $sql) or die(mysqli_error($db));
+
+  $commentsql = "SELECT users_comments.body, users.username FROM users_comments, users WHERE post_id = $postid AND users_comments.user_id = users.User_ID";
+  $commentresult = mysqli_query($db, $commentsql) or die(mysqli_error($db));
+
+  while ($commentrow = mysqli_fetch_array($commentresult)) {
+    $comments .= "<b>".$commentrow[1]."</b>: ".$commentrow[0]."</br>";
+  }
+
   if (mysqli_num_rows($result2) < 1) {
     $posts .= "<div class='jumbotron'>".$row[1]."<br><img src='assets/imgs/users/".$row[4]."' width=100 height=100 /> <br> <br><b>" .$row[3]."</b>: ".$row[2]."<hr>
               <form action='index.php?&postid=".$row[0]."' method='post'>
                 <input type='submit' name='like' value='Like'>
               </form>
               Likes: " .$row[5]."
+
+              <form action='index.php?postid=".$row[0]."' method='post'>
+                <textarea  class='form-control' name='commentbody' rows='3' cols='40'></textarea>
+                <input type='submit' name='comment' value='Comment!' class='btn btn-light'>
+              </form>"
+
+              .$comments."
 
     </div></br>";
   } else {
@@ -57,6 +84,13 @@ while ($row = mysqli_fetch_array($result)) {
                 <input type='submit' name='like' value='Unlike'>
               </form>
               Likes: " .$row[5]."
+
+              <form action='index.php?postid=".$row[0]."' method='post'>
+                <textarea  class='form-control' name='commentbody' rows='3' cols='40'></textarea>
+                <input type='submit' name='comment' value='Comment!' class='btn btn-light'>
+              </form>"
+
+              .$comments."
 
     </div></br>";
   }
